@@ -4,7 +4,11 @@
   "description": "React powered component renderer",
   "main": "rts.js",
   "dependencies": {
-    "react": "^0.13.3"
+    "exports-loader": "^0.6.2",
+    "imports-loader": "^0.6.4",
+    "react": "^0.14.0",
+    "react-dom": "^0.14.0",
+    "whatwg-fetch": "^0.10.0"
   },
   "devDependencies": {
     "babel-core": "^5.8.25",
