@@ -1,48 +1,59 @@
-import React from 'react';
+import React, { PropTypes } from 'react';
+import ClientSideHelpers from '../../src/ClientSideHelpers.es6'
 
 class NameBox extends React.Component {
   constructor(props) {
     super(props);
-    this.state = { name: this.props.name };
+    this.state = {
+      name: this.props.name,
+      suggestions: []
+    };
 
-    // http://stackoverflow.com/questions/29577977/react-ref-and-setstate-not-working-with-es6
-    this._handleNameChange = this._handleNameChange.bind(this);
+    this._handleChooseSelection = this._handleChooseSelection.bind(this);
   }
 
   render() {
-    const name = this.state.name;
-    const message = name ? <p>Hi {name}, how splendid!</p> : '';
     return (
       <div>
-        <form method="post" action="" onSubmit={this._handleNameChange}>
-          <input type="text" name="name" defaultValue={name} placeholder="e.g. Bob" />
-          <input type="submit" value="Change name" />
+        <form ref="form" action="/name" method="post" className="form-inline" onSubmit={ClientSideHelpers.handleFormSubmitClientSide}>
+          <input name="name" autoComplete="off" type="text" className="form-control" value={this.state.name} onChange={this._handleFieldChange.bind(this)} placeholder="e.g. Bob" />
+          <button type="submit" className="btn btn-default">Change Name</button>
         </form>
-        <p>I haven't yet implemented <a href="https://facebook.github.io/flux/docs/overview.html#structure-and-data-flow">Flux</a> or an equivalent, so changing your name in this NameBox won't propagate that change to other components.</p>
-        <p>The idea is that a client-side only 'dispatcher' collects state change events, alters the state and then selectively re-renders using <code>renderInto</code>. This logic would be duplicated by the fallback no-JS ruby endpoints this client-side JS overrides, so it must be kept simple.</p>
+        <div className="">
+          <ul className="">
+            {this.state.suggestions.map(suggestedName => {
+              return <li key={suggestedName}><a href="" onClick={this._handleChooseSelection}>{suggestedName}</a></li>
+            })}
+          </ul>
+        </div>
       </div>
     );
   }
 
   // Handlers
 
-  _handleNameChange(event) {
+  _handleFieldChange(event) {
+    const partial = event.target.value;
+    this.setState({ name: partial });
+
+    fetch('/api/suggestions/names?partial=' + partial)
+      .then(res => res.json())
+      .then(json => {
+        this.setState({ suggestions: json });
+      });
+  }
+
+  _handleChooseSelection(event) {
     event.preventDefault();
-    // this.setState({ name: event.target.value });
-    const e = new CustomEvent(
-      'rerender', {
-        detail: {
-          action: 'nameChange',
-          name: event.target.name.value
-        }
-      }
-    );
-    window.dispatchEvent(e);
+    this.setState({ name: event.target.text }, () => {
+      ClientSideHelpers.handleFormSubmitClientSide(this.refs.form, false);
+    });
   }
 }
 
 NameBox.propTypes = {
-  name: React.PropTypes.string
+  name: PropTypes.string,
+  suggestions: PropTypes.arrayOf(PropTypes.string)
 };
 
 export default NameBox;
